Increase statistical confidence in 'oracle' distribution
One hypothesis from @rrtoledo about the incorrect number of retries in
the "large" case is that the oracle skews the distribution in such a
way we still have "bad wheels" in this case. We tried to increase the
size and confidence interval of the test and it still makes sense, but
worth investigating more.