Using generics inputs (#139)
## Content

So far, we could only use Element as input type. This was fixed to a
byte array of a specific size for the whole repository.

This PR aims at supporting mroe input types, and allowing the use of
different input types across examples and tests.

To do so, I have created a `ToBytes` trait and updating the relevant
struct and functions to use a generic type implementing it. The hash
functions then only have to call `to_be_bytes()` that is the one
function defined in the trait.

I have also written some macros to implement `ToBytes`, and so support,
some standard types.

## Pre-submit checklist

- Branch
    - [x] Tests are provided (if possible)
    - [x] Commit sequence broadly makes sense
    - [x] Key commits have useful messages
- PR
    - [x] No clippy warnings in the CI
    - [x] Self-reviewed the diff
    - [x] Useful pull request description
    - [x] Reviewer requested
- Documentation
    - [x] Update README file (if relevant)
    - [x] Update documentation website (if relevant)

## Comments

We should be able to support both BLS and KES (with current Cardano
params) signatures.

## Issue(s)

Relates to #30